+++ src/renderer/components/ImageCropper/index.tsx
@@ -13,6 +13,8 @@ import Dialog, {
   FooterActionButton,
 } from '../Dialog'
 
+import Icon from '../Icon'
+
 import useTranslationFunction from '../../hooks/useTranslationFunction'
 import { LastUsedSlot, rememberLastUsedPath } from '../../utils/lastUsedPaths'
 
@@ -46,6 +48,9 @@ export default function ImageCropper({
   const shade = useRef<HTMLImageElement>(null)
   // a wrapper inside which we drag an image
   const container = useRef<HTMLImageElement>(null)
+  // a temporary canvas for rotation and result output
+  // so we don't create a new canvas element on each operation
+  const tmpCanvas = useRef<HTMLCanvasElement>(null)
 
   const dragging = useRef<boolean>(false)
   // where we start dragging
@@ -61,13 +66,20 @@ export default function ImageCropper({
   const containerScale = useRef<number>(1.0)
   const zoom = useRef<number>(1.0)
   const initialZoom = useRef<number>(1.0)
+  /** we use radians */
+  const rotation = useRef<number>(0.0)
 
   const onSubmit = () => {
     if (!fullImage.current) {
       return
     }
 
-    const canvas = document.createElement('canvas')
+    const canvas = tmpCanvas.current
+
+    if (!canvas) {
+      return
+    }
+
     canvas.width = targetWidth.current
     canvas.height = targetHeight.current
 
@@ -103,11 +115,71 @@ export default function ImageCropper({
     })()
   }
 
+  const onZoomIn = () => {
+    zoom.current += 0.01
+    moveImages(posX.current, posY.current)
+  }
+
+  const onZoomOut = () => {
+    zoom.current -= 0.01
+    moveImages(posX.current, posY.current)
+  }
+
   const onZoomReset = () => {
+    if (!cutImage.current || !fullImage.current) {
+      return
+    }
     posX.current = 0
     posY.current = 0
     zoom.current = initialZoom.current
     moveImages(0, 0)
+    if (rotation.current > 0) {
+      rotation.current = 0
+      cutImage.current.src = filepath
+      fullImage.current.src = filepath
+    }
+  }
+
+  const onRotateImages = () => {
+    if (!tmpCanvas.current || !fullImage.current || !cutImage.current) {
+      return
+    }
+
+    const canvas = tmpCanvas.current
+
+    const ctx = canvas.getContext('2d')
+
+    if (!ctx) {
+      throw new Error('Somehow canvas context for image rotation went missing')
+    }
+
+    canvas.width = fullImage.current.clientHeight
+    canvas.height = fullImage.current.clientWidth
+
+    rotation.current += Math.PI * 0.5
+
+    // FIXME: we now only assume Math.PI * 0.5 rotations
+    if (rotation.current > Math.PI * 1.5) {
+      rotation.current = 0
+    }
+
+    ctx.translate(canvas.width / 2, canvas.height / 2)
+    ctx.rotate(Math.PI * 0.5)
+    ctx.drawImage(
+      fullImage.current as CanvasImageSource,
+      fullImage.current.clientWidth / -2,
+      fullImage.current.clientHeight / -2
+    )
+
+    const dataURL = canvas.toDataURL()
+
+    cutImage.current.src = dataURL
+    fullImage.current.src = dataURL
+
+    setupImages()
+
+    // NOTE: we swap positions here
+    moveImages(posY.current, posX.current)
   }
 
   const makeSelector = (x: number, y: number) => {
@@ -289,27 +361,50 @@ export default function ImageCropper({
               src={filepath}
             />
           </div>
-          <div className={styles.imageCropperHint}>
-            {tx('image_cropper_hint_desktop')}
+          <div className={styles.imageCropperControls}>
+            <button
+              className={styles.imageCropperControlsButton}
+              onClick={onZoomIn}
+              aria-label={tx('menu_zoom_in')}
+            >
+              <Icon coloring='navbar' icon='plus' size={18} />
+            </button>
+            <button
+              className={styles.imageCropperControlsButton}
+              onClick={onZoomOut}
+              aria-label={tx('menu_zoom_out')}
+            >
+              <Icon coloring='navbar' icon='minus' size={18} />
+            </button>
+            <button
+              className={styles.imageCropperControlsButton}
+              onClick={onRotateImages}
+              aria-label={tx('ImageEditorHud_rotate')}
+            >
+              <Icon coloring='navbar' icon='rotate-right' size={24} />
+            </button>
           </div>
+          <canvas ref={tmpCanvas} style={{ display: 'none' }}></canvas>
         </DialogContent>
       </DialogBody>
       <DialogFooter>
-        <FooterActions>
-          <FooterActionButton
-            onClick={() => {
-              onCancel()
-              onClose()
-            }}
-          >
-            {tx('cancel')}
-          </FooterActionButton>
-          <FooterActionButton onClick={onZoomReset}>
+        <FooterActions align='spaceBetween'>
+          <FooterActionButton onClick={onZoomReset} aria-label={tx('reset')}>
             {tx('reset')}
           </FooterActionButton>
-          <FooterActionButton onClick={onSubmit}>
-            {tx('save')}
-          </FooterActionButton>
+          <FooterActions>
+            <FooterActionButton
+              onClick={() => {
+                onCancel()
+                onClose()
+              }}
+            >
+              {tx('cancel')}
+            </FooterActionButton>
+            <FooterActionButton onClick={onSubmit}>
+              {tx('save')}
+            </FooterActionButton>
+          </FooterActions>
         </FooterActions>
       </DialogFooter>
     </Dialog>