+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## Added
 - More shortcuts to switch between chats: `Ctrl + PageDown`, `Ctrl + PageUp`, `Ctrl + Tab`, `Ctrl + Shift + Tab` #3984
+- add on-screen controls to ImageCropper and the ability to rotate by 90 degrees #3893
+- add on-screen controls to ImageCropper and the ability to rotate by 90 degrees #3893
 
 ### Changed
 - reword advanced setting "Disable Background Sync For All Accounts" -> "Only Synchronize the Currently Selected Account" #3960